Refactor duplicated snake_case logic to utils and fix login prompt
