+++ src/lib.rs
@@ -22,6 +22,7 @@ pub mod repl;
 pub mod stdlib;
 pub mod transpiler;
 pub mod typechecker;
+pub mod utils;
 pub mod version;
 pub mod wfl_config;
